+++ .idea/.gitignore
@@ -6,3 +6,4 @@
 # Datasource local storage ignored files
 /dataSources/
 /dataSources.local.xml
+/**/dataSources.xml
