+++ apps/server/src/services/alt-reminder/index.ts
@@ -1,1 +1,1 @@
-﻿export * from "./service";
+﻿export * from "./discord-controller";
