+++ apps/server/src/services/managed-message/index.ts
@@ -1,1 +1,3 @@
 export * from "./service";
+export * from "./discord-controller";
+export * from "./http-controller";
