+++ apps/server/src/services/poll/index.ts
@@ -1,1 +1,2 @@
 export * from "./service";
+export * from "./discord-controller";
